refactor: move all physics derived quantities to their own module
